server/requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
redis[hiredis]
fastapi-limiter